temp commit. HMake recompiled using header-units.

+++ hconfigure/header/BasicTargets.hpp
@@ -78,7 +78,7 @@ struct BTarget // BTarget
     virtual void updateBTarget(unsigned short round, class Builder &builder);
     virtual void printMutexLockRoutine(unsigned short round);
     virtual void preSort(Builder &builder, unsigned short round);
-    virtual void duringSort(Builder &builder, unsigned short round, unsigned short indexInTopologicalSortComparator);
+    virtual void duringSort(Builder &builder, unsigned short round, unsigned int indexInTopologicalSortComparator);
 };
 bool operator<(const BTarget &lhs, const BTarget &rhs);
